requirements.txt
python-dotenv
pydantic


# For Langchain
langchain
langchain-openai
langchain-community
langchain-chroma
langchain-postgres
chroma-hnswlib
chromadb


# For MCP
fastmcp

# misc
tqdm
pandas
openpyxl
xlsxwriter

# sqlite
aiosqlite